+++ unittest/pbdata/reads/RegionAnnotations_gtest.cpp
@@ -161,21 +161,21 @@ TEST(RegionAnnotationsTest, SubreadIntervals_2)
     RegionAnnotations ras(HOLENUMBER, regions, TYPES);
 
     std::vector<ReadInterval> ris = ras.SubreadIntervals(WHOLE_LENGTH, true, false);
-    EXPECT_EQ(ris.size(), 1);  // (112, WHOLE_LENGTH, -1)
+    EXPECT_EQ(ris.size(), 1u);  // (112, WHOLE_LENGTH, -1)
     EXPECT_EQ(ris[0].start, 112);
-    EXPECT_EQ(ris[0].end, WHOLE_LENGTH);
+    EXPECT_EQ(ris[0].end, static_cast<int>(WHOLE_LENGTH));
 
     // no insert, no hq && require adapter, require hq
     ris = ras.SubreadIntervals(WHOLE_LENGTH, true, true);
-    EXPECT_EQ(ris.size(), 0);
+    EXPECT_EQ(ris.size(), 0u);
 
     // no require adapter, no require hq
     ris = ras.SubreadIntervals(WHOLE_LENGTH, false, false);  // no insert
-    EXPECT_EQ(ris.size(), 0);
+    EXPECT_EQ(ris.size(), 0u);
 
     // no require adapter, require hq
     ris = ras.SubreadIntervals(WHOLE_LENGTH, false, true);  // no hq
-    EXPECT_EQ(ris.size(), 0);
+    EXPECT_EQ(ris.size(), 0u);
 }
 
 TEST(RegionAnnotationsTest, SubreadIntervals_3)
@@ -188,18 +188,18 @@ TEST(RegionAnnotationsTest, SubreadIntervals_3)
 
     // require adapter, no require hq
     std::vector<ReadInterval> ris = ras.SubreadIntervals(WHOLE_LENGTH, true, false);
-    EXPECT_EQ(ris.size(), 0);
+    EXPECT_EQ(ris.size(), 0u);
 
     // require adapter, require hq
     ris = ras.SubreadIntervals(WHOLE_LENGTH, true, true);
-    EXPECT_EQ(ris.size(), 0);
+    EXPECT_EQ(ris.size(), 0u);
 
     // no require adapter, no require hq
     ris = ras.SubreadIntervals(WHOLE_LENGTH, false, false);
-    EXPECT_EQ(ris.size(), 1);
+    EXPECT_EQ(ris.size(), 1u);
     EXPECT_EQ(ris[0], ReadInterval(0, 170, -1));
 
     // no require adapter, require hq
     ris = ras.SubreadIntervals(WHOLE_LENGTH, false, true);
-    EXPECT_EQ(ris.size(), 0);
+    EXPECT_EQ(ris.size(), 0u);
 }
